# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "4.3.0",
    "title": "Job's Portal",
    "description": "You get best here",
    "license": {
      "name": "Sukhman",
      "url": "https://epitometechnologies.org/"
    }
  },
  "host": "localhost:8000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/user/signup": {
      "post": {
        "tags": [
          "Users API"
        ],
        "summary": "Sign up",
        "description": "Create new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "type": "object",
            "description": "Insert User data",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "Users API"
        ],
        "summary": "Login",
        "description": "login  in system",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "type": "object",
            "description": "Login User ",
            "schema": {
              "$ref":"#/definitions/Login"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Logged In"
          }
        }
      }
    },
    "/user/forget_pass": {
      "post": {
        "tags": [
          "Users API"
        ],
        "summary": "Forget Password",
        "description": "forget pass of User",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "type": "object",
            "description": "Forget Pass",
            "schema": {
              "$ref":"#/definitions/Forget_password"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "token sent  Succesfully"
          }
        }
      }
    },
    "/user/reset_pass/{id}/{token}": {
      "post": {
        "tags": [
          "Users API"
        ],
        "summary": "Reset Password",
        "description": "Reset pass of User",
        "parameters": [
          {
            "name": "id",
            "in": "path", 
            "description": "Insert Id "
          },
          {
            "name": "token",
            "in": "path", 
            "description": "Insert token"
          },
          {
            "name": "User",
            "in": "body",
            "type": "object",
            "description": "Insert new Password ",
            "schema" : {
              "$ref":"#/definitions/Reset_password"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": { 
            "description":"Password update Succesfully"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "User data with given ID",
        "tags": [
          "Users API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of user that we want to find",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "escription": "Users data"
          }
        }
      },
      "patch": {        
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of user that we want to update",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ],
        "summary": "User data with given ID",
        "tags": [
          "Users API"
        ],
        "responses": {
          "200": {
            "description": "User data"
            
          }
        }
      },
      "delete": {
        "tags": [
          "Users API"
        ],
        "summary": "delete user data with given ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of user that we want to delete",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User data"
          }
        }
      }
     },
...
    "/user/singlejob/{id}": {
      "get": {
        "summary": " Single Job  detail",
        "tags": [
          "Users API"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of Job that we want to find",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "single Job data"
          }
        }
      
    }
  },
  "/user/job/org/{id}": {
    "get": {
      "summary": "All Jobs from one ORG",
      "tags": [
        "Users API"
      ],
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of ORG to find All Jobs",
          "type": "string"
        }
      ],
      "responses": {
        "200": {
          "description": "ORG Jobs"
        }
      }
    
  }
},
...
  },
  "definitions": {
    "User": {
      "properties": {
        "FirstName": {
          "type": "string"
        },
        "LastName": {
          "type": "string"
        },
        "Email": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        },
        "ConfirmPassword": {
          "type": "string"
        },
        "Phone": {
          "type": "number"
        },
        "Country": {
          "type": "string"
        },
        "Skills": {
          "type": "object",
          "properties":{
            "Professional_Skills": {
              "type": "array",
          "items": {
            "type": "string"
          }
            },
            "Training_Intership": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "Education": {
          "type": "array",
          "items": {
          "type": "object",
          "properties":{
            "Tenth": {
              "type": "string"
            }
          } 
         }
        },
        "CurrentCompany": {
          "type": "string"
        }
      }
    },
    "Login": {
      "properties":{
        "Email":{
          "type": "string"
        },
        "Password":{
          "type": "string"
        }
      }
    },
    "Forget_password": {
      "properties":{
        "Email":{
          "type": "string"
        }
      }
    },
    "Reset_password": {
      "properties":{
        "Email":{
          "type": "string"
        },
        "Password":{
          "type": "string"
        },
        "ConfirmPassword":{
          "type": "string"
        }
      }
    },
    "UpdateUser": {
      "properties": {
        "FirstName": {
          "type": "string"
        },
        "LastName": {
          "type": "string"
        },
        "Phone": {
          "type": "number"
        },
        "Country": {
          "type": "string"
        },
        "Skills": {
          "type": "object",
          "properties":{
            "Professional_Skills": {
              "type": "array",
          "items": {
            "type": "string"
          }
            },
            "Training_Intership": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "Education": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "CurrentCompany": {
          "type": "string"
        }
      }
    },
...
  }
  }